requirements.txt
fastapi>=0.110.0
httpx>=0.27.0
pydantic>=2.0.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
langgraph>=0.2.0
radon>=6.0.0
pylint>=3.0.0
